requirements.txt
pyyaml>=6.0
python-dateutil>=2.9
markdownify>=0.14
pdfplumber>=0.11
python-pptx>=1.0
openpyxl>=3.1
fastapi>=0.115
uvicorn>=0.34
msal>=1.28
requests>=2.31
